Fix remaining sign-conversion warnings in tests and examples
Fixes GCC 15 -Werror=conversion warnings in test suites and VulkanSCTriangleExample:

Test Files:
- DeviceObjectManagerTests.cpp:69 - Cast count_if result to size_t
- MemoryPoolManagerTests.cpp:65 - Cast count_if result to size_t

VulkanSCTriangleExample.cpp:
- Line 445: Cast fileSize to streamsize for file.read()
- Line 752: Cast frameCount and elapsed to float for FPS calculation

All warnings now resolved. Library builds cleanly on Windows MSVC and Linux GCC 15.

+++ tests/vulkansc/MemoryPoolManagerTests.cpp
@@ -62,8 +62,8 @@ public:
     }
 
     void printSummary() const {
-        size_t passed = count_if(results.begin(), results.end(),
-                                [](const TestResult& r) { return r.passed; });
+        size_t passed = static_cast<size_t>(count_if(results.begin(), results.end(),
+                                                     [](const TestResult& r) { return r.passed; }));
         size_t failed = results.size() - passed;
 
         cout << "\n=============================================================================\n";
